Trait sc_consensus_grandpa_rpc::GrandpaApiServer

pub trait GrandpaApiServer<Notification, Hash, Number>: Sized + Send + Sync + 'static {
    // Required methods
    fn round_state<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<ReportedRoundStates>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn prove_finality<'life0, 'async_trait>(
        &'life0 self,
        block: Number
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<EncodedFinalityProof>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe_justifications(
        &self,
        subscription_sink: SubscriptionSink
    ) -> SubscriptionResult;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where Notification: Send + Sync + 'static + Serialize,
             Hash: Send + Sync + 'static,
             Number: Send + Sync + 'static + DeserializeOwned { ... }
}

Server trait implementation for the GrandpaApi RPC API.

Required Methods

fn round_state<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<ReportedRoundStates>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the state of the current best round state as well as the ongoing background rounds.

fn prove_finality<'life0, 'async_trait>( &'life0 self, block: Number ) -> Pin<Box<dyn Future<Output = RpcResult<Option<EncodedFinalityProof>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Prove finality for the given block number by returning the Justification for the last block in the set and all the intermediary headers to link them together.

fn subscribe_justifications( &self, subscription_sink: SubscriptionSink ) -> SubscriptionResult

Returns the block most recently finalized by Grandpa, alongside side its justification.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where Notification: Send + Sync + 'static + Serialize, Hash: Send + Sync + 'static, Number: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.